requirements.txt
# AI Ethics Testing Framework Requirements

# Core web framework
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.0.0

# Database drivers
mysqlclient>=2.2.0

# Authentication
PyJWT>=2.8.0

# HTTP requests and utilities
requests>=2.28.0
python-dotenv>=1.0.0

# AI API SDKs
openai>=1.12.0
anthropic>=0.25.0
google-generativeai>=0.3.0
cohere>=4.47.0

# Production dependencies
supervisor>=4.2.0

# Required for full functionality
python-dateutil>=2.8.0

# Optional dependencies for enhanced functionality
# numpy>=1.24.0  # For advanced statistical analysis
# pandas>=2.0.0  # For data manipulation
# matplotlib>=3.7.0  # For additional plotting
# seaborn>=0.12.0  # For statistical visualizations

# Development dependencies (uncomment for development)
# pytest>=7.0.0  # For testing
# pylint>=2.17.0  # For code quality
# black>=23.0.0  # For code formatting
# requests>=2.31.0  # For API testing
